requirements.txt
numpy
onnxruntime
Pillow
opencv-python-headless
flask 